Support Dual Stack Listen on all interfaces
When 0.0.0.0 or [::] are specified as the server IP Address, i.e.
'Listen on all interfaces', listen on both IPv6 and IPv4.

Previously when listening on all interfaces, clients which resolve the
IPv6 address of the server would have connections refused.

Requires setting to IPAddress.IPv6Any and setting the IPv6Only socket
option to false to allow IPv4 compatibility.